refactor(1807): rename reyn.plugins → reyn.gateway (clean-break)
"plugins" was a misnomer — the package is an inbound webhook GATEWAY (the
sample_line / sample_slack webhook handlers), not a plugin system. Rename to
reyn.gateway aligns the name with intent (and with the "gateway layer" naming
in the #1812 concept doc).

Clean-break, no shim (owner policy): git mv (byte-identical, rename-detection)
+ atomic repoint of every importer. src/reyn/plugins → src/reyn/gateway and the
mirror tests/plugins → tests/gateway; all `reyn.plugins` dotted refs (imports,
docstring entry-point examples, path strings) → `reyn.gateway`.

Straggler 0 verified with scripts/verify_package_move.py (the mechanize tool's
first production use): `verify_package_move.py reyn.plugins` reports 0 residual
across dotted-import / dotted-literal / segment-path / bare-source-path. Manual
cross-grep confirms. Repo-wide collection: 7586 tests, no import breakage.

The docs/plugins/ directory + any prose "plugins layer" mentions are a
docs-maintainer follow-up (per the #311 playbook); the `reyn.plugins` dotted
refs inside docs are repointed here for code-correctness.

Closes #1807.
